move noble set up to noble client
