+++ e2e/src/utils/mod.rs
@@ -1,13 +1,12 @@
 pub mod authorization;
 pub mod base_account;
 pub mod ethereum;
-pub mod file;
 pub mod hyperlane;
 pub mod ibc;
 pub mod icq;
 pub mod manager;
-pub mod noble;
 pub mod osmosis;
+pub mod parse;
 pub mod persistence;
 pub mod polytone;
 pub mod processor;
